Introduces a GATT client example
This example compliments the ble-gatt-server example by providing a client that is able to connect to it. Assuming that ble-gatt-server is running on another device, once this client connects it will display anything entered into the server. The example demonstrates best practices by running a foreground service that is able to automatically re-connect and display notifications to an encrypted characteristic. It also demonstrates how these notifications can manifest themselves through an Android binding.