Fix issue with null BluetoothDevice field.
The exception was thrown on reading device field which was null.
The property hasn't been assigned, because callbacks were invoked
in unforeseen order. To fix the issue assigning device was moved
to before() callback and to avoid resuming twice the same coroutine
the resume() method has been moved to done() callback.