Trigger service discovery if interrupted in `onConnectionUpdated`
It appears that when a service discovery is in progress and
`onConnectionUpdated` is called, the service discovery callback
`onServicesDiscovered` is never invoked. By checking to see
`serviceDiscoverRequested` is set, we can restart the service discovery
and get a result back.

This may potentially fix #536, but I was seeing this behaviour
consistently when connecting to another BLE device